requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
discord.py==1.2.3
gspread==3.1.0 
lxml==4.4.1
mecab-python3==0.996.2
numpy==1.17.2
oauth2client==4.1.3
requests==2.22.0
requests-oauthlib==1.2.0
youtube-dl==2019.10.16
